utilities/meshutils/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
async-timeout==4.0.3
    # via bleak
bleak==0.21.1
    # via -r requirements.in
build==1.2.1
    # via pip-tools
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
dbus-fast==2.21.1
    # via bleak
idna==3.6
    # via requests
packaging==24.0
    # via build
pip-tools==7.4.1
    # via -r requirements.in
pycryptodomex==3.20.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pyserial==3.5
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.10.0
    # via bleak
urllib3==2.2.1
    # via requests
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
